+++ R/ggmid_mid_conditional.R
@@ -8,10 +8,10 @@
 #' @param type a character string specifying the type of the plot. One of "iceplot" or "centered". If "centered", the ICE values of each observation are set to zero at the leftmost point of the varriable.
 #' @param theme a character string specifying the color theme or any item that can be used to define "color.theme" object.
 #' @param term an optional character string specifying an interaction term. If passed, the ICE curve for the specified term is plotted.
-#' @param var.alpha a name of the variable to be used to set \code{alpha}.
-#' @param var.color a name of the variable to be used to set \code{colour}.
-#' @param var.linetype a name of the variable to be used to set \code{linetype}.
-#' @param var.linewidth a name of the variable to be used to set \code{linewidth}.
+#' @param var.alpha a name of the variable or an expression to be used to set \code{alpha}.
+#' @param var.color a name of the variable or an expression to be used to set \code{colour}.
+#' @param var.linetype a name of the variable or an expression to be used to set \code{linetype}.
+#' @param var.linewidth a name of the variable or an expression to be used to set \code{linewidth}.
 #' @param dots logical. If \code{TRUE}, the points representing the predictions for each observation are plotted.
 #' @param sample an optional vector specifying the names of observations to be plotted.
 #' @param ... optional parameters to be passed to the main layer.
@@ -29,7 +29,6 @@ ggmid.mid.conditional <- function(
     object, type = c("iceplot", "centered"), theme = NULL, term = NULL,
     var.alpha = NULL, var.color = NULL, var.linetype = NULL, var.linewidth = NULL,
     dots = TRUE, sample = NULL, ...) {
-  cl <- match.call(expand.dots = TRUE)
   type <- match.arg(type)
   theme <- color.theme(theme)
   use.theme <- inherits(theme, "color.theme")
@@ -61,30 +60,38 @@ ggmid.mid.conditional <- function(
   pl <- ggplot2::ggplot(
     data = obs, ggplot2::aes(x = .data[[variable]], y = .data[[yvar]])
   )
-  if (!is.null(cl$var.alpha)) {
-    alp <- characterize(cl$var.alpha)
-    pl <- pl +
-      ggplot2::aes(alpha = .data[[alp]]) +
+  if (!is.null(alp <- substitute(var.alpha))) {
+    if (is.character(alp))
+      alp <- str2lang(alp)
+    obs$.alp <- eval(alp, envir = obs)
+    con$.alp <- eval(alp, envir = con)
+    pl <- pl + ggplot2::aes(alpha = .data[[".alp"]]) +
       ggplot2::labs(alpha = alp)
   }
-  if (set.color <- !is.null(cl$var.color)) {
-    col <- characterize(cl$var.color)
-    pl <- pl +
-      ggplot2::aes(colour = .data[[col]]) +
+  if (set.color <- !is.null(col <- substitute(var.color))) {
+    if (is.character(col))
+      col <- str2lang(col)
+    obs$.col <- eval(col, envir = obs)
+    con$.col <- eval(col, envir = con)
+    pl <- pl + ggplot2::aes(colour = .data[[".col"]]) +
       ggplot2::labs(colour = col)
   }
-  if (!is.null(cl$var.linetype)) {
-    lty <- characterize(cl$var.linetype)
-    pl <- pl +
-      ggplot2::aes(linetype = .data[[lty]]) +
+  if (!is.null(lty <- substitute(var.linetype))) {
+    if (is.character(lty))
+      lty <- str2lang(lty)
+    obs$.lty <- eval(lty, envir = obs)
+    con$.lty <- eval(lty, envir = con)
+    pl <- pl + ggplot2::aes(linetype = .data[[".lty"]]) +
       ggplot2::labs(linetype = lty)
     if (!is.discrete(obs[, lty]))
       pl <- pl + ggplot2::scale_linetype_binned()
   }
-  if (!is.null(cl$var.linewidth)) {
-    lwd <- characterize(cl$var.linewidth)
-    pl <- pl +
-      ggplot2::aes(linewidth = .data[[lwd]]) +
+  if (!is.null(lwd <- substitute(var.linewidth))) {
+    if (is.character(lwd))
+      lwd <- str2lang(lwd)
+    obs$.lwd <- eval(lwd, envir = obs)
+    con$.lwd <- eval(lwd, envir = con)
+    pl <- pl + ggplot2::aes(linewidth = .data[[".lwd"]]) +
       ggplot2::labs(linewidth = lwd)
     if (is.discrete(obs[, lwd])) {
       pl <- pl + ggplot2::scale_linewidth_discrete(range = c(0, 1))
@@ -95,13 +102,13 @@ ggmid.mid.conditional <- function(
   pl <- pl + ggplot2::geom_line(
     data = con, mapping = ggplot2::aes(group = .data[[".id"]]), ...)
   if (dots) {
-    pl <- pl + ggplot2::geom_point()
+    pl <- pl + ggplot2::geom_point(data = obs)
   }
   if (set.color) {
     if (!use.theme)
       theme <- "bluescale"
     pl <- pl + scale_color_theme(theme = theme,
-                                 discrete = is.discrete(obs[[col]]))
+                                 discrete = is.discrete(obs$.col))
   }
   pl
 }