+++ NEWS.md
@@ -1,8 +1,14 @@
-# midr 0.5.0 (TBR)
+## midr 0.5.0 (TBR)
 
-First release on CRAN.
+[First release on CRAN.]
 
-# midr 0.4.3-0.4.8.900
+## midr 0.4.3.903
+
+-   Update `interpret()` to ensure that the value of `link` argument should be a character string.
+
+-   Update `ggmid.mid.conditional()` and `plot.mid.conditional()`: `var.color` and the similar arguments can take an expression as input.
+
+## midr 0.4.3-0.4.8.902
 
 midr is in the release process. We will submit the package to CRAN by mid-January 2025 .